Move as much as possible to rover module
