+++ kifi-backend/modules/common/app/com/keepit/common/store/ProdStoreModule.scala
@@ -3,7 +3,6 @@ package com.keepit.common.store
 import java.io.File
 
 import com.amazonaws.services.s3.transfer.TransferManager
-import com.keepit.rover.sensitivity.{ PornWordLikelihood, InMemoryPornWordLikelihoodStore, S3PornWordLikelihoodStore, PornWordLikelihoodStore }
 import com.keepit.search.tracking.{ InMemoryProbablisticLRUStoreImpl, S3ProbablisticLRUStoreImpl, ProbablisticLRUStore }
 import org.apache.commons.io.FileUtils
 import play.api.Play.current
@@ -65,12 +64,6 @@ trait ProdStoreModule extends StoreModule {
   }
 
   @Provides @Singleton
-  def bayesPornDetectorStore(amazonS3Client: AmazonS3, accessLog: AccessLog): PornWordLikelihoodStore = {
-    val bucketName = S3Bucket(current.configuration.getString("amazon.s3.bayes.porn.detector.bucket").get)
-    new S3PornWordLikelihoodStore(bucketName, amazonS3Client, accessLog)
-  }
-
-  @Provides @Singleton
   def s3ImageConfig: S3ImageConfig = {
     val bucket = current.configuration.getString("cdn.bucket")
     val base = current.configuration.getString("cdn.base")
@@ -112,15 +105,6 @@ abstract class DevStoreModule[T <: ProdStoreModule](override val prodStoreModule
     whenConfigured("amazon.s3.install.bucket")(
       prodStoreModule.kifiInstallationStore(amazonS3ClientProvider.get, accessLog)
     ).getOrElse(new InMemoryKifiInstallationStoreImpl())
-
-  @Provides @Singleton
-  def bayesPornDetectorStore(amazonS3ClientProvider: Provider[AmazonS3], accessLog: AccessLog) = {
-    whenConfigured("amazon.s3.bayes.porn.detector.bucket")(
-      prodStoreModule.bayesPornDetectorStore(amazonS3ClientProvider.get, accessLog)
-    ).getOrElse(new InMemoryPornWordLikelihoodStore() {
-        override def syncGet(key: String) = Some(PornWordLikelihood(Map("a" -> 1f)))
-      })
-  }
 
   @Provides @Singleton
   def s3ImageConfig: S3ImageConfig =
