+++ kifi-backend/modules/common/app/com/keepit/rover/article/content/ArticleContent.scala
@@ -2,6 +2,7 @@ package com.keepit.rover.article.content
 
 import com.keepit.rover.article.Article
 import com.keepit.rover.fetcher.HttpRedirect
+import com.keepit.rover.model.HttpProxy
 import com.kifi.macros.json
 import org.joda.time.DateTime
 
@@ -27,6 +28,20 @@ case class HttpInfo(
   redirects: Seq[HttpRedirect],
   httpContentType: Option[String],
   httpOriginalContentCharset: Option[String])
+
+object HttpInfo {
+
+  def ok: HttpInfo = {
+    HttpInfo(
+      200,
+      Some("OK"),
+      Seq(),
+      None,
+      None
+    )
+  }
+
+}
 
 trait NormalizationInfoHolder {
   def normalization: NormalizationInfo
