+++ kifi-backend/modules/rover/app/com/keepit/rover/fetcher/HttpFetcherModule.scala
