+++ kifi-backend/modules/rover/app/com/keepit/rover/fetcher/apache/ApacheFetchResponse.scala
